Moved domain adaptation to mixing
