+++ albumentations/augmentations/__init__.py
@@ -2,8 +2,6 @@ from .blur.functional import *
 from .blur.transforms import *
 from .crops.functional import *
 from .crops.transforms import *
-from .domain_adaptation.functional import *
-from .domain_adaptation.transforms import *
 from .dropout.channel_dropout import *
 from .dropout.coarse_dropout import *
 from .dropout.functional import *
@@ -15,6 +13,8 @@ from .geometric.functional import *
 from .geometric.resize import *
 from .geometric.rotate import *
 from .geometric.transforms import *
+from .mixing.domain_adaptation import *
+from .mixing.domain_adaptation_functional import *
 from .mixing.functional import *
 from .mixing.transforms import *
 from .spectrogram.transform import *
