--- albumentations/augmentations/domain_adaptation/__init__.py
@@ -1,2 +1,0 @@
-from .functional import *
-from .transforms import *
